refactor: move applyMessages into its own file
